QS: make dynamic tiles configurable (1/2)
instead of using a bunch of provider settings var
like it was on 4.3 and almost all other roms are doing
use one and split like we do for the normal tiles. This
is more consistent and saves us to observe a bunch of values
As well this is integrated into the one layout screen we have. So
all QS related is easy and logical usable for the user on one
screen

Change-Id: Ic1b8d021f4bf418257df268262b7eff9474d4819

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettingsController.java